// packages/engine/src/battlecast/types/monster.ts
export interface Creature {
  id: string;
  name: string;
  displayName: string;
  monsterData: MonsterData;
  team: 'red' | 'blue';
// ...
}